+++ Resources/Private/Php/update.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+composer update --with-dependencies
+composer dumpautoload --optimize
+find . -name ".git" -exec rm -rf {} \;