+++ ext_localconf.php
@@ -95,6 +95,14 @@ $boot = static function () {
     <INCLUDE_TYPOSCRIPT: source="FILE:EXT:seminars/Configuration/TSconfig/ContentElementWizard.txt">
 '
     );
+
+    // register the time slot wizard
+    // @deprecated will be removed in seminars 4.0
+    $GLOBALS['TYPO3_CONF_VARS']['SYS']['formEngine']['nodeRegistry'][1558632705] = [
+        'nodeName' => 'time_slot_wizard',
+        'priority' => 70,
+        'class' => \OliverKlee\Seminars\BackEnd\TimeSlotWizard::class,
+    ];
 };
 
 $boot();