+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Add a BE forms wizard for creating a series of time slots (#820)
 - Add PHPStan to the CI builds (#763)
 - Allow installations up to PHP 8.0 (#758)